Added a WallMover Actor component and a Interactor Scene component
Wall Mover:-
A component that when put upon any object takes in a move offset and a time to move the object there along with a condition. When the condition is satisfied the move begins.
Interactor:-
When pressing F returns scene actors that Block the collision channel Interact.
Also created a blueprint node that returns the hit results of the interact action.
Created BP_Gargoyle which starts shining when interacted with(for now, later it's gonna be used as a switch to open doors using the mover component.
Also switched to Visual Studio Code because Visual studio 2019 was legging back in IntelliSense and I like having IntelliSense