The switch system is complete
Added a switch system to the WallMover Component.
The player can now interact with any objects that are derived from the Switch class and block the collision channel called interactions.
Any scene object with a WallMover component on it can take in objects derived from the switch class and once all of the switches are interacted with (activated) then the WallMover class will execute its movement code.